pyproject.toml
[tool.poetry]
name = "Template FastAPI"
version = "0.1.0"
description = ""
authors = ["Evgeny Aleshin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"


[tool.poetry.group.dev.dependencies]
pyyaml = "^6.0.1"
httpx = "^0.26.0"
pytest = "^8.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
